Refactor repository imports and update DataSource interface methods
