+++ src/main/kotlin/data/datasource/DataSource.kt
@@ -1,7 +1,7 @@
-package org.baghdad.data.storage
+package org.baghdad.data.datasource
 
 interface DataSource<T> {
     fun loadAll(): List<T>
-    fun saveAll(items: List<T>)
+    fun update(items: List<T>)
     fun append(item: T)
 }
