+++ src/main/kotlin/logic/repositories/ProjectRepository.kt
@@ -1,6 +1,6 @@
 package org.baghdad.logic.repositories
 
-import org.baghdad.logic.entities.ProjectEntity
+import org.baghdad.logic.model.entities.ProjectEntity
 
 interface ProjectRepository {
