+++ src/main/kotlin/logic/repositories/TaskRepository.kt
@@ -1,12 +1,12 @@
 package org.baghdad.logic.repositories
 
-import org.baghdad.logic.entities.TaskEntity
+import org.baghdad.logic.model.entities.TaskEntity
 
 interface TaskRepository {
 
     fun createTask(task: TaskEntity): Boolean
     fun getTaskById(id: String): TaskEntity
-    fun getTasksByProjectId(projectId: String): List<TaskEntity>
+    fun getTasksByProjectId(id: String): List<TaskEntity>
     fun updateTask(task: TaskEntity): Boolean
     fun deleteTask(id: String): Boolean
 }
