+++ src/main/kotlin/logic/repositories/UserRepository.kt
@@ -1,10 +1,10 @@
 package org.baghdad.logic.repositories
 
-import org.baghdad.logic.entities.UserEntity
+import org.baghdad.logic.model.entities.UserEntity
 
 interface UserRepository {
 
-    fun createUser(user: UserEntity)
+    fun createUser(username: String, userCreator: UserEntity)
     fun getUserByUsername(username: String): UserEntity?
     fun getAllUsers(): List<UserEntity>
 }
